Revert to two-state casino
While originally playing around with the examples I added a 'stupid'
state to the casino, but now I want the two-state version back... but
with the fixes to the makefile and state id labeling (see code). Would
be better if I had put this in git before making any changes, I know.