Fix problem with Baum Welch
The parameter outputTable was set to "no" in the backward algorithm
block that was doing baum welch estimation... but we needed the damn dp
table to be saved. Oddly enough, it was set to "no" in the original
hmmoc release, too. I'll need to investigate to figure out whether (if)
it was working then.